+++ dataQueue.h
@@ -4,10 +4,10 @@
 #define QUEUE_SIZE 20
 
 // Queues [data] onto queue
-void enqueue(int data);
+void enqueue(double data);
 
 // Pops off [data] from queue
-int dequeue(void);
+double dequeue(void);
 
 // Returns whether the sampled data exceeds [LIGHT_INTENSITY_THRESHOLD]
 int doesExceedThreshold(void);
